[CITE-163] Addressed PR comments
